Add counts
