+++ front/app/modules/commercial/insights/utils/getSelectedCategoryFilter.ts
@@ -1,0 +1,8 @@
+const getSelectedCategoryFilter = (categoryQuery?: string) =>
+  categoryQuery
+    ? 'category'
+    : categoryQuery === ''
+    ? 'notCategorized'
+    : 'allInput';
+
+export default getSelectedCategoryFilter;
